Add package setup and CLI interface
- Add pyproject.toml for package configuration
- Add CLI module with corefud_scorer and ua_scorer commands
- Add INSTALL.md with installation and usage guide
- Add .gitignore for Python project
- Add MANIFEST.in for package data
- Update scorer_main.py with full implementation
